modify 화주 페이지 대시보드 기능 연결 (idrp-116))
오더(status), 후기(해당 오더 건에 대한 차주 대상), 결제(status) 대시보드(통합) 기능 연결

+++ backend/backend_idle/src/main/java/com/fullstack/service/ReviewServiceImpl.java
@@ -31,7 +31,7 @@ public class ReviewServiceImpl implements ReviewService {
                 .orElseThrow(() -> new IllegalArgumentException("작성자 정보를 찾을 수 없습니다."));
 
         // 대상(차주) 정보 조회
-        CustomerEntity target = customerRepository.findById(requestDto.getTargetId())
+        CustomerEntity target = customerRepository.findById(requestDto.getTargetId().intValue())
                 .orElseThrow(() -> new IllegalArgumentException("리뷰 대상 정보를 찾을 수 없습니다."));
 
         // TODO: 비즈니스 로직 추가
